Refactor: Combine project creation steps and types into a single configuration file
Related to #1422
