--- src/app/constants/projectCreationSteps.ts
@@ -1,8 +1,0 @@
-export const projectCreationSteps = [
-  "Type",
-  "Details",
-  "Socials",
-  "Logo",
-  "Media", // Display only if ProjectType is null or EscrowPayment.
-  "Affiliates"
-  ];
